Fixed Bug

+++ src/AppBundle/Service/TabsService.php
@@ -729,9 +729,11 @@ class TabsService extends BaseService
                 // Delete TaskToCheckListItem entries
                 $deleteTaskToChecklistItemID = $this->entityManager->getRepository('AppBundle:Taskstochecklistitems')->DeleteDuplicateChecklistItems($task->getTaskid(),$checkListItem['ChecklistItemID']);
                 foreach ($deleteTaskToChecklistItemID as $item) {
-                    $removeEntry = $this->entityManager->getRepository('AppBundle:Taskstochecklistitems')->find($item['TaskToChecklistItemID']);
-                    $this->entityManager->remove($removeEntry);
-                    $this->entityManager->flush();
+                    $removeEntry = $this->entityManager->getRepository('AppBundle:Taskstochecklistitems')->find((int)$item['TaskToChecklistItemID']);
+                    if ($removeEntry) {
+                        $this->entityManager->remove($removeEntry);
+                        $this->entityManager->flush();
+                    }
                 }
 
                 // Empty res2 and create all entries
